refact: add double cast

+++ OpenSeesMatlab/+post/+resp/ShellRespStepData.m
@@ -73,7 +73,7 @@ classdef ShellRespStepData < post.resp.ResponseBase
 
 
 
-            obj.eleTags = eleTags(:).';
+            obj.eleTags = double(eleTags(:).');
 
             obj.eleClassCache = containers.Map('KeyType','double','ValueType','int32');
             obj.eleNGPCache   = containers.Map('KeyType','double','ValueType','int32');
@@ -95,7 +95,7 @@ classdef ShellRespStepData < post.resp.ResponseBase
             if nargin < 2 || isempty(eleTags)
                 eleTags = obj.eleTags;
             end
-            eleTags = eleTags(:).';
+            eleTags = double(eleTags(:).');
             nEle    = numel(eleTags);
 
             [maxGP, maxFib] = shell_fill_caches(obj, eleTags, obj.mex_);
